moves get_current_user from yools.utils into libs.auth.jwt_utils
