+++ openvair/libs/auth/jwt_utils.py
@@ -1,0 +1,56 @@
+"""Need to write"""
+from typing import Dict
+
+import jwt
+from fastapi import Depends, HTTPException, security
+
+from openvair import config
+from openvair.libs.log import get_logger
+
+oauth2schema = security.OAuth2PasswordBearer('/auth/')
+JWT_SECRET = config.data['jwt'].get('secret')
+TOKEN_TYPE = config.data['jwt'].get('token_type')
+ALGORITHM = config.data['jwt'].get('algorithm')
+ACCESS_TOKEN_EXPIRE_MINUTES = config.data['jwt'].get(
+    'access_token_expiration_minutes', 30
+)
+REFRESH_TOKEN_EXPIRATION_DAYS = config.data['jwt'].get(
+    'refresh_token_expiration_days', 7
+)
+LOG = get_logger(__name__)
+
+
+def get_current_user(token: str = Depends(oauth2schema)) -> Dict:
+    """Retrieves the current user from a JWT token.
+
+    Args:
+        token (str): The JWT token to decode.
+
+    Returns:
+        Dict: The user object decoded from the token.
+
+    Raises:
+        HTTPException: If the token is invalid or expired.
+    """
+    try:
+        payload: Dict = jwt.decode(
+            token,
+            JWT_SECRET,
+            algorithms=[
+                ALGORITHM,
+            ],
+        )
+
+        if payload['type'] != 'access':
+            raise HTTPException(status_code=401, detail='Invalid access token')
+
+        payload.update({'token': token})
+        LOG.info(f'User authenticated: {payload["username"]}')
+    except jwt.ExpiredSignatureError:
+        LOG.error('Token expired')
+        raise HTTPException(status_code=401, detail='Token expired')
+    except jwt.InvalidTokenError:
+        LOG.error('Invalid token')
+        raise HTTPException(status_code=401, detail='Invalid token')
+    else:
+        return payload
