+++ openvair/modules/block_device/entrypoints/api.py
@@ -31,7 +31,7 @@ from fastapi import Depends, APIRouter, status
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import get_current_user
+from openvair.libs.auth.jwt_utils import get_current_user
 from openvair.modules.block_device.entrypoints import schemas
 from openvair.modules.block_device.entrypoints.crud import InterfaceCrud
