+++ openvair/modules/image/entrypoints/api.py
@@ -36,7 +36,7 @@ from starlette.concurrency import run_in_threadpool
 
 from openvair.config import TMP_DIR
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import get_current_user
+from openvair.libs.auth.jwt_utils import get_current_user
 from openvair.modules.image.config import CHUNK_SIZE
 from openvair.modules.image.entrypoints import schemas, exceptions
 from openvair.modules.image.entrypoints.crud import ImageCrud
