+++ openvair/modules/network/entrypoints/api.py
@@ -23,7 +23,7 @@ from fastapi_pagination import Page, paginate
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import get_current_user
+from openvair.libs.auth.jwt_utils import get_current_user
 from openvair.modules.network.entrypoints import schemas
 from openvair.modules.network.entrypoints.crud import InterfaceCrud
