+++ openvair/modules/tools/utils.py
@@ -39,7 +39,7 @@ from contextlib import contextmanager
 
 import jwt
 import xmltodict
-from fastapi import Depends, HTTPException, security
+from fastapi import HTTPException, security
 from sqlalchemy.exc import OperationalError
 
 from openvair import config
@@ -196,42 +196,6 @@ def create_tokens(user: Dict) -> Dict:
         'refresh_token': refresh_token,
         'token_type': TOKEN_TYPE,
     }
-
-
-def get_current_user(token: str = Depends(oauth2schema)) -> Dict:
-    """Retrieves the current user from a JWT token.
-
-    Args:
-        token (str): The JWT token to decode.
-
-    Returns:
-        Dict: The user object decoded from the token.
-
-    Raises:
-        HTTPException: If the token is invalid or expired.
-    """
-    try:
-        payload: Dict = jwt.decode(
-            token,
-            JWT_SECRET,
-            algorithms=[
-                ALGORITHM,
-            ],
-        )
-
-        if payload['type'] != 'access':
-            raise HTTPException(status_code=401, detail='Invalid access token')
-
-        payload.update({'token': token})
-        LOG.info(f'User authenticated: {payload["username"]}')
-    except jwt.ExpiredSignatureError:
-        LOG.error('Token expired')
-        raise HTTPException(status_code=401, detail='Token expired')
-    except jwt.InvalidTokenError:
-        LOG.error('Invalid token')
-        raise HTTPException(status_code=401, detail='Invalid token')
-    else:
-        return payload
 
 
 def get_block_devices_info() -> List[Dict[str, str]]:
