+++ openvair/modules/virtual_machines/entrypoints/api.py
@@ -37,7 +37,7 @@ from fastapi_pagination import Page, paginate
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import get_current_user
+from openvair.libs.auth.jwt_utils import get_current_user
 from openvair.modules.virtual_machines.entrypoints import schemas
 from openvair.modules.virtual_machines.entrypoints.crud import VMCrud
