+++ openvair/modules/virtual_network/entrypoints/api.py
@@ -41,7 +41,7 @@ from fastapi.responses import JSONResponse
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import get_current_user
+from openvair.libs.auth.jwt_utils import get_current_user
 from openvair.modules.virtual_network.entrypoints import schemas
 from openvair.modules.virtual_network.entrypoints.crud import VirtualNetworkCrud
