+++ openvair/modules/volume/entrypoints/api.py
@@ -28,7 +28,7 @@ from fastapi_pagination import Page, paginate
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import get_current_user
+from openvair.libs.auth.jwt_utils import get_current_user
 from openvair.libs.validation.validators import Validator
 from openvair.modules.volume.entrypoints import schemas
 from openvair.modules.volume.entrypoints.crud import VolumeCrud
